5_HandsOn_Tutorials/5.8_Message_Queues/5.8.3_NATS/01_basic_setup/requirements.txt
nats-py==2.6.0
